Change generation settings to be blank
Currently, functionality relies on the hidden surface not being generated. If the surface is generated -- through the editor, for example -- water and other ores may be present, preventing saving of ores or allowing other ores to be created. Change generation to not generate water, entities, or decoratives. Cliffs are still generated but are removed when ores are added to the hidden surface.